[GF:2467352:BrazilPerforce://brazil/src/shared/platform/IonJava/release]

+++ src/com/amazon/ion/impl/IonTokenReader.java
@@ -384,7 +384,7 @@ public class IonTokenReader
         this.in = new IonCharacterReader( r );
     }
 
-    public int getConsumedAmount() {
+    public long getConsumedAmount() {
         return in.getConsumedAmount();
     }
